# docs/swagger.yaml
basePath: /api
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.Course:
    properties:
      category:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: integer
      lessons:
        items:
          $ref: '#/definitions/models.Lesson'
        type: array
      teacher_id:
        type: integer
      title:
        type: string
      updatedAt:
        type: string
    required:
    - title
    type: object
  models.Lesson:
    properties:
      content:
        type: string
      course_id:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      order:
        type: integer
      title:
        type: string
      updatedAt:
        type: string
    required:
    - title
    type: object
host: localhost:8090
info:
  contact: {}
  description: Uzaktan Eğitim Platformu REST API
  title: GoLearn API
  version: "1.0"
paths:
  /auth/login:
    post:
      parameters:
      - description: Giriş Bilgileri
        in: body
        name: body
        required: true
        schema: {}
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Kullanıcı Girişi
      tags:
      - Auth
  /auth/register:
    post:
      parameters:
      - description: Kayıt Bilgileri
        in: body
        name: body
        required: true
        schema: {}
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      summary: Kullanıcı Kaydı
      tags:
      - Auth
  /courses:
    get:
      parameters:
      - description: Sayfa
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Kategori
        in: query
        name: category
        type: string
      - description: Sıralama (asc, desc)
        in: query
        name: sort
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Course'
            type: array
      summary: Kursları Filtrele ve Listele
      tags:
      - Course
    post:
      responses: {}
      security:
      - BearerAuth: []
      summary: Yeni Kurs Oluştur
      tags:
      - Course
  /courses/{id}:
    delete:
      responses: {}
      security:
      - BearerAuth: []
      summary: Kurs Sil
      tags:
      - Course
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Course'
      summary: Kurs Detayı
      tags:
      - Course
    put:
      responses: {}
      security:
      - BearerAuth: []
      summary: Kurs Güncelle
      tags:
      - Course
  /courses/{id}/lessons:
    get:
      responses: {}
      summary: Kursa Ait Dersleri Listele
      tags:
      - Lesson
    post:
      responses: {}
      security:
      - BearerAuth: []
      summary: Kursa Yeni Ders Ekle
      tags:
      - Lesson
  /lessons/{id}/quiz:
    get:
      responses: {}
      summary: Derse Ait Quiz Getir
      tags:
      - Quiz
    post:
      responses: {}
      security:
      - BearerAuth: []
      summary: Yeni Quiz Oluştur
      tags:
      - Quiz
  /quiz/{id}/submit:
    post:
      responses: {}
      security:
      - BearerAuth: []
      summary: Quiz Çöz ve Puanla
      tags:
      - Quiz
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"